+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### v3.9.x - [Unreleased]
 
+#### Fixed
+
+- MySQL databases backup when using remote MySQL server
+
 ### v3.11.4 - 2020-01-17
 
 #### Fixed